R: move some tests into more specific namespaces
