+++ src/System.CommandLine.Tests/Binding/TypeWithInvokeAndCtor.cs
@@ -2,8 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace System.CommandLine.Tests.Binding
