+++ src/System.CommandLine.Tests/Invocation/CancelOnProcessTerminationTests.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using Xunit;
 
-namespace System.CommandLine.Tests
+namespace System.CommandLine.Tests.Invocation
 {
     public class CancelOnProcessTerminationTests
     {
