+++ src/System.CommandLine.Tests/Invocation/CommandHandlerTests.cs
@@ -2,25 +2,35 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.CommandLine.Invocation;
-using System.CommandLine.Binding;
-using System.CommandLine.Tests.Binding;
-using FluentAssertions;
-using System.Linq;
 using System.Threading.Tasks;
+using FluentAssertions;
 using Xunit;
 
-namespace System.CommandLine.Tests
+namespace System.CommandLine.Tests.Invocation
 {
-    public class BindingCommandHandlerTests
+    public class CommandHandlerTests
     {
         private readonly TestConsole _console = new TestConsole();
+
+        [Fact]
+        public async Task Specific_invocation_behavior_can_be_specified_in_the_command()
+        {
+            var wasCalled = false;
+
+            var command = new Command("command");
+            command.Handler = CommandHandler.Create(() => wasCalled = true);
+
+            var parser = new Parser(command);
+
+            await parser.InvokeAsync("command", _console);
+
+            wasCalled.Should().BeTrue();
+        }
 
         [Fact]
         public async Task Method_parameters_on_the_invoked_method_are_bound_to_matching_option_names()
         {
             var wasCalled = false;
-            const string commandLine = "command --age 425 --name Gandalf";
-            object[] expectedArgumets = { "Gandalf", 425 };
 
             void Execute(string name, int age)
             {
@@ -36,13 +46,9 @@ namespace System.CommandLine.Tests
             command.AddOption(
                 new Option("--age",
                            argument: new Argument<int>()));
-            var handler = CommandHandler.Create<string, int>(Execute);
-            command.Handler = handler;
+            command.Handler = CommandHandler.Create<string, int>(Execute);
 
-            var arguments = handler.Binder.GetInvocationArguments(command.CreateBindingContext(commandLine));
-            arguments.Should().BeEquivalentSequenceTo(expectedArgumets);
-
-            await command.InvokeAsync(commandLine, _console);
+            await command.InvokeAsync("command --age 425 --name Gandalf", _console);
 
             wasCalled.Should().BeTrue();
         }
@@ -51,8 +57,6 @@ namespace System.CommandLine.Tests
         public async Task Method_parameters_on_the_invoked_method_can_be_bound_to_hyphenated_option_names()
         {
             var wasCalled = false;
-            const string commandLine = "command --first-name Gandalf";
-            object[] expectedArgumets = { "Gandalf" };
 
             void Execute(string firstName)
             {
@@ -63,13 +67,9 @@ namespace System.CommandLine.Tests
             var command = new Command("command");
             command.AddOption(new Option("--first-name",
                                          argument: new Argument { Arity = ArgumentArity.ExactlyOne }));
-            var handler = CommandHandler.Create<string>(Execute);
-            command.Handler = handler;
+            command.Handler = CommandHandler.Create<string>(Execute);
 
-            var arguments = handler.Binder.GetInvocationArguments(command.CreateBindingContext(commandLine));
-            arguments.Should().BeEquivalentSequenceTo(expectedArgumets);
-
-            await command.InvokeAsync(commandLine, _console);
+            await command.InvokeAsync("command --first-name Gandalf", _console);
 
             wasCalled.Should().BeTrue();
         }
@@ -78,8 +78,6 @@ namespace System.CommandLine.Tests
         public async Task Method_parameters_on_the_invoked_method_can_be_bound_to_option_names_case_insensitively()
         {
             var wasCalled = false;
-            const string commandLine = "command --age 425 --NAME Gandalf";
-            object[] expectedArgumets = { "Gandalf", 425 };
 
             void Execute(string name, int Age)
             {
@@ -91,23 +89,17 @@ namespace System.CommandLine.Tests
             var command = new Command("command");
             command.AddOption(new Option("--NAME", argument: new Argument { Arity = ArgumentArity.ExactlyOne }));
             command.AddOption(new Option("--age", argument: new Argument<int>()));
-            var handler = CommandHandler.Create<string, int>(Execute);
-            command.Handler = handler;
+            command.Handler = CommandHandler.Create<string, int>(Execute);
 
-            var arguments = handler.Binder.GetInvocationArguments(command.CreateBindingContext(commandLine));
-            arguments.Should().BeEquivalentSequenceTo(expectedArgumets);
-
-            await command.InvokeAsync(commandLine, _console);
+            await command.InvokeAsync("command --age 425 --NAME Gandalf", _console);
 
             wasCalled.Should().BeTrue();
         }
 
-        [Fact(Skip = "Waiting for fix for nulls in BeEquivalentSequenceTo")]
+        [Fact]
         public async Task Method_parameters_on_the_invoked_method_do_not_need_to_be_matched()
         {
             var wasCalled = false;
-            const string commandLine = "command";
-            object[] expectedArgumets = { null, 0 };
 
             void Execute(string name, int age)
             {
@@ -119,13 +111,9 @@ namespace System.CommandLine.Tests
             var command = new Command("command");
             command.AddOption(new Option("--name", argument: new Argument<string>() { Arity = ArgumentArity.ExactlyOne }));
             command.AddOption(new Option("--age", argument: new Argument<int>()));
-            var handler = CommandHandler.Create<string, int>(Execute);
-            command.Handler = handler;
+            command.Handler = CommandHandler.Create<string, int>(Execute);
 
-            var arguments = handler.Binder.GetInvocationArguments(command.CreateBindingContext(commandLine));
-            arguments.Should().BeEquivalentSequenceTo(expectedArgumets);
-
-            await command.InvokeAsync(commandLine, _console);
+            await command.InvokeAsync("command", _console);
 
             wasCalled.Should().BeTrue();
         }
@@ -134,8 +122,6 @@ namespace System.CommandLine.Tests
         public async Task Method_parameters_on_the_invoked_method_can_be_bound_to_option_names_by_alias()
         {
             var wasCalled = false;
-            const string commandLine = "command -a 425 -n Gandalf";
-            object[] expectedArgumets = { "Gandalf", 425 };
 
             void Execute(string name, int Age)
             {
@@ -147,13 +133,9 @@ namespace System.CommandLine.Tests
             var command = new Command("command");
             command.AddOption(new Option(new[] { "-n", "--NAME" }, argument: new Argument { Arity = ArgumentArity.ExactlyOne }));
             command.AddOption(new Option(new[] { "-a", "--age" }, argument: new Argument<int>()));
-            var handler = CommandHandler.Create<string, int>(Execute);
-            command.Handler = handler;
+            command.Handler = CommandHandler.Create<string, int>(Execute);
 
-            var arguments = handler.Binder.GetInvocationArguments(command.CreateBindingContext(commandLine));
-            arguments.Should().BeEquivalentSequenceTo(expectedArgumets);
-
-            await command.InvokeAsync(commandLine, _console);
+            await command.InvokeAsync("command -a 425 -n Gandalf", _console);
 
             wasCalled.Should().BeTrue();
         }
@@ -162,8 +144,6 @@ namespace System.CommandLine.Tests
         public async Task Method_parameters_on_the_invoked_lambda_are_bound_to_matching_option_names()
         {
             var wasCalled = false;
-            const string commandLine = "command --age 425 --name Gandalf";
-            object[] expectedArgumets = { "Gandalf", 425 };
 
             var command = new Command("command");
             command.AddOption(new Option("--name", "", new Argument<string>()));
@@ -175,12 +155,7 @@ namespace System.CommandLine.Tests
                 age.Should().Be(425);
             });
 
-            var handler = command.Handler as ReflectionCommandHandler;
-
-            var arguments = handler.Binder.GetInvocationArguments(command.CreateBindingContext(commandLine));
-            arguments.Should().BeEquivalentSequenceTo(expectedArgumets);
-
-            await command.InvokeAsync(commandLine, _console);
+            await command.InvokeAsync("command --age 425 --name Gandalf", _console);
 
             wasCalled.Should().BeTrue();
         }
@@ -189,24 +164,16 @@ namespace System.CommandLine.Tests
         public async Task Method_parameters_of_type_ParseResult_receive_the_current_ParseResult_instance()
         {
             var wasCalled = false;
-            const string commandLine = "command -x 123";
 
             var command = new Command("command");
             command.AddOption(new Option("-x", "", new Argument<int>()));
             command.Handler = CommandHandler.Create<ParseResult>(result =>
-            {
-                wasCalled = true;
-                result.ValueForOption("-x").Should().Be(123);
-            });
+                               {
+                                   wasCalled = true;
+                                   result.ValueForOption("-x").Should().Be(123);
+                               });
 
-            var handler = command.Handler as ReflectionCommandHandler;
-
-            var arguments = handler.Binder.GetInvocationArguments(command.CreateBindingContext(commandLine));
-            arguments.Count().Should().Be(1);
-            arguments.First().Should().NotBeNull();
-            arguments.First().Should().BeOfType<ParseResult>();
-
-            await command.InvokeAsync(commandLine, _console);
+            await command.InvokeAsync("command -x 123", _console);
 
             wasCalled.Should().BeTrue();
         }
@@ -215,7 +182,6 @@ namespace System.CommandLine.Tests
         public async Task Method_parameters_of_type_IConsole_receive_the_current_console_instance()
         {
             var wasCalled = false;
-            const string commandLine = "command";
 
             var command = new Command("command");
             command.AddOption(new Option("-x", "", new Argument<int>()));
@@ -225,17 +191,28 @@ namespace System.CommandLine.Tests
                 console.Out.Write("Hello!");
             });
 
-            var handler = (ReflectionCommandHandler) command.Handler;
-
-            var arguments = handler.Binder.GetInvocationArguments(command.CreateBindingContext(commandLine));
-
-            arguments.Should()
-                     .ContainSingle(a => a is IConsole);
-           
-            await command.InvokeAsync(commandLine, _console);
+            await command.InvokeAsync("command", _console);
 
             wasCalled.Should().BeTrue();
             _console.Out.ToString().Should().Be("Hello!");
+        }
+
+        [Fact]
+        public async Task Method_parameters_of_type_InvocationContext_receive_the_current_InvocationContext_instance()
+        {
+            var wasCalled = false;
+
+            var command = new Command("command");
+            command.AddOption(new Option("-x", "", new Argument<int>()));
+            command.Handler = CommandHandler.Create<InvocationContext>(context =>
+            {
+                wasCalled = true;
+                context.ParseResult.ValueForOption("-x").Should().Be(123);
+            });
+
+            await command.InvokeAsync("command -x 123", _console);
+
+            wasCalled.Should().BeTrue();
         }
     }
 }
