+++ src/System.CommandLine.Tests/Invocation/InvocationPipelineTests.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using Xunit;
 
-namespace System.CommandLine.Tests
+namespace System.CommandLine.Tests.Invocation
 {
     public class InvocationPipelineTests
     {
