+++ src/System.CommandLine.Tests/MethodBinderTests.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.CommandLine.Builder;
 using System.CommandLine.Invocation;
-using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
