+++ src/System.CommandLine/Binding/IBinder.cs
@@ -7,4 +7,14 @@ namespace System.CommandLine.Binding
     {
         object GetTarget(BindingContext context);
     }
+
+
+
+
+
+
+
+
+
+
 }
